Move ClientConfigType out of reporting plugin
