+++ package.json
@@ -604,6 +604,7 @@
     "@kbn/reporting-export-types-png-common": "link:packages/kbn-reporting/export_types/png_common",
     "@kbn/reporting-mocks-server": "link:packages/kbn-reporting/mocks_server",
     "@kbn/reporting-plugin": "link:x-pack/plugins/reporting",
+    "@kbn/reporting-public": "link:packages/kbn-reporting/public",
     "@kbn/reporting-server": "link:packages/kbn-reporting/server",
     "@kbn/resizable-layout": "link:packages/kbn-resizable-layout",
     "@kbn/resizable-layout-examples-plugin": "link:examples/resizable_layout_examples",
