+++ packages/kbn-reporting/public/README.md
@@ -1,0 +1,3 @@
+# @kbn/reporting-public
+
+Empty package generated by @kbn/generate
