+++ packages/kbn-reporting/public/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@kbn/reporting-public",
+  "private": true,
+  "version": "1.0.0",
+  "license": "SSPL-1.0 OR Elastic License 2.0"
+}
