+++ tsconfig.base.json
@@ -1192,6 +1192,8 @@
       "@kbn/reporting-mocks-server/*": ["packages/kbn-reporting/mocks_server/*"],
       "@kbn/reporting-plugin": ["x-pack/plugins/reporting"],
       "@kbn/reporting-plugin/*": ["x-pack/plugins/reporting/*"],
+      "@kbn/reporting-public": ["packages/kbn-reporting/public"],
+      "@kbn/reporting-public/*": ["packages/kbn-reporting/public/*"],
       "@kbn/reporting-server": ["packages/kbn-reporting/server"],
       "@kbn/reporting-server/*": ["packages/kbn-reporting/server/*"],
       "@kbn/resizable-layout": ["packages/kbn-resizable-layout"],
