+++ x-pack/plugins/reporting/public/management/components/report_diagnostic.tsx
@@ -19,8 +19,8 @@ import {
   EuiText,
   EuiTitle,
 } from '@elastic/eui';
+import { ClientConfigType } from '@kbn/reporting-public';
 import { ReportingAPIClient, DiagnoseResponse } from '../../lib/reporting_api_client';
-import { ClientConfigType } from '../../plugin';
 
 interface Props {
   apiClient: ReportingAPIClient;
